Add files via upload
File can take a while to open, trust me it's not crashing! several minutes! 

I won't explain the answer for day 2 suffice to say once you find a consistent pattern theres a simple linear formula y=mx+b where x is the generation and y is the sum of the indecies for the plants alive during that generation.